// src/types/SEOType.d.ts
/**
 * SEO Props
 */
export type SEOProps = {
  /**
   * Title of The Page
   * @type {string}
   */
  title: string;
  /**
   * Description of the page
   * @type {string}
   */
  description: string;
  /**
   * Content of the page or OG
   * @type {string}
   */
  content: string;
  /**
   * Icon 57 of Apple Icon
   * @type {string}
   */
  icon57: string;

  /**
   * Icon 72 of Apple Icon
   * @type {string}
   */
  icon72: string;
  /**
   * OG Description for FB
   * @type {string}
   */
  ogDescription: string;
  /**
   * OG Image standard
   * @type {string}
   */
  ogImage: string;
};